Refactored timeAgo function from CommentCard into separate util file
